chore(seed): add seed script
