+++ lib/index.ts
@@ -1,3 +1,2 @@
-export * from './constants';
 export * from './validators';
 export { default as prisma } from './prisma';
