+++ lib/seed.ts
@@ -1,4 +1,48 @@
-import { prisma, SEED_THEMES, SEED_LANGUAGES, SEED_ROLES } from '@/lib';
+const { PrismaClient } = require('@prisma/client');
+
+const SEED_LANGUAGES = [
+  {
+    name: 'English',
+    code: 'en-US',
+  },
+  {
+    name: 'English',
+    code: 'en-GB',
+  },
+  {
+    name: 'Spanish',
+    code: 'es-ES',
+  },
+  {
+    name: 'Portuguese',
+    code: 'pt-BR',
+  },
+  {
+    name: 'French',
+    code: 'fr-CA',
+  },
+  {
+    name: 'Chinese',
+    code: 'zh-TW',
+  },
+];
+const SEED_ROLES = [
+  {
+    name: 'participant',
+    description: 'Research pariticpant',
+  },
+  {
+    name: 'user',
+    description: 'Researcher',
+  },
+  {
+    name: 'admin',
+    description: 'User with all priviledges',
+  },
+];
+const SEED_THEMES = ['dark', 'light'];
+
+const prisma = new PrismaClient();
 
 const seedThemes = async () => {
   for (const theme of SEED_THEMES) {
